Fixed issue with links not being found
Google recently changed the way they present the image data, and so the links were no longer being scraped.
I figured out how to get the image urls with the new system and made the appropriate changes so it would work. 

Unfortunately, google no longer provides file format data so I had to try and retrieve it from the url of the image, which does not work in some cases.